+++ plugins/Goals/tests/UI/GoalsPages_spec.js
@@ -8,8 +8,6 @@
  */
 
 describe("GoalsPages", function () {
-  this.timeout(0);
-
   var generalParams = 'idSite=1&period=year&date=2012-08-09',
     urlBaseGeneric = 'module=CoreHome&action=index&',
     urlBase = urlBaseGeneric + generalParams;