+++ plugins/Morpheus/tests/UI/Morpheus_spec.js
@@ -6,8 +6,6 @@
  */
 
 describe("Morpheus", function () {
-    this.timeout(0);
-
     var url = "?module=Morpheus&action=demo";
 
     before(function () {